# backend/api/openapi.yaml
openapi: "3.1.0"
info: {title: Oh-My-AIHub API, version: "0.1.0", description: "Oh-My-AIHub 前后端契约的唯一来源（见 ADR）。覆盖全部 /api/** JSON 接口；/v1/**、/v1beta/** 外部模型协议入口只登记路径与认证。\n\n约定：金额一律为十进制字符串；时间为 RFC 3339；错误统一为 {\"error\":{\"code\",\"message\"}}；除 GET/HEAD/OPTIONS 外的 /api 请求须携带与 Host 同源的 Origin。\n\n扩展字段 x-access 声明路由门禁，由契约测试与路由表逐项对照：public 无需会话；session 需会话（允许首次改密前访问）；ready 需会话且已完成首次改密；admin 在 ready 基础上需管理员；gateway_key 凭网关 API Key。"}
servers:
  - url: /
    description: "同源部署：开发环境由 Vite 代理，Compose 环境由 Nginx 代理。"
...
paths:
...
  /api/wallet:
    get:
      operationId: getWallet
      tags: [wallet]
      summary: "我的钱包"
      x-access: ready
      security:
        - sessionCookie: []
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WalletResponse"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
  /api/wallet/entries:
    get:
      operationId: listWalletEntries
      tags: [wallet]
      summary: "我的账本分录"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - name: limit
          in: query
          required: false
          description: "返回条数。"
          schema: {type: integer, default: 20}
        - name: before
          in: query
          required: false
          description: "分录 ID 游标，返回早于该分录的记录。"
          schema: {type: string, pattern: "^[0-9]+$"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LedgerEntryPage"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '422': {$ref: "#/components/responses/Unprocessable"}
...
  /api/admin/ledger/accounts/{accountID}/wallet:
    get:
      operationId: getAdminLedgerAccountWallet
      tags: [admin-ledger]
      summary: "用户账户钱包"
      x-access: admin
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/accountID"
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WalletEnvelope"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
  /api/admin/ledger/accounts/{accountID}/entries:
    get:
      operationId: listAdminLedgerAccountEntries
      tags: [admin-ledger]
      summary: "用户账户账本分录"
      x-access: admin
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/accountID"
        - name: limit
          in: query
          required: false
          description: "返回条数。"
          schema: {type: integer, default: 20}
        - name: before
          in: query
          required: false
          description: "分录 ID 游标。"
          schema: {type: string, pattern: "^[0-9]+$"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LedgerEntryPage"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/admin/ledger/system-accounts/{systemKind}/wallet:
    get:
      operationId: getAdminLedgerSystemWallet
      tags: [admin-ledger]
      summary: "系统账户钱包"
      x-access: admin
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/systemKind"
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WalletEnvelope"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
  /api/admin/ledger/system-accounts/{systemKind}/entries:
    get:
      operationId: listAdminLedgerSystemEntries
      tags: [admin-ledger]
      summary: "系统账户账本分录"
      x-access: admin
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/systemKind"
        - name: limit
          in: query
          required: false
          description: "返回条数。"
          schema: {type: integer, default: 20}
        - name: before
          in: query
          required: false
          description: "分录 ID 游标。"
          schema: {type: string, pattern: "^[0-9]+$"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LedgerEntryPage"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/admin/ledger/adjustments:
    post:
      operationId: createLedgerAdjustment
      tags: [admin-ledger]
      summary: "管理员账本调整"
      x-access: admin
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Idempotency-Key"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LedgerAdjustmentRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LedgerTransactionEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/admin/ledger/bad-debts:
    post:
      operationId: createBadDebtTransfer
      tags: [admin-ledger]
      summary: "坏账划转"
      x-access: admin
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Idempotency-Key"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/BadDebtRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LedgerTransactionEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
...
  /api/c2c/me:
    get:
      operationId: getC2CMyActivity
      tags: [c2c]
      summary: "我的订单与交易"
      x-access: ready
      security:
        - sessionCookie: []
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/C2CMyActivity"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: oma_session
      description: "登录后由服务端设置的 HttpOnly 会话 Cookie。"
    bearerKey: {type: http, scheme: bearer, description: "平台 API Key：Authorization: Bearer <key>（OpenAI 协议）。"}
    anthropicKey:
      type: apiKey
      in: header
      name: x-api-key
      description: "平台 API Key（Anthropic 协议）。"
    geminiKey:
      type: apiKey
      in: header
      name: x-goog-api-key
      description: "平台 API Key（Gemini 协议）。"
  parameters:
    accountID:
      name: accountID
      in: path
      required: true
      description: "账户 ID（UUID）。"
      schema: {type: string, format: uuid}
...
    systemKind:
      name: systemKind
      in: path
      required: true
      description: "系统账户类型。"
      schema: {type: string, enum: [platform_incentive, platform_loss]}
    Idempotency-Key:
      name: Idempotency-Key
      in: header
      required: false
      description: "幂等键；相同键重复提交返回同一结果。"
      schema: {type: string}
...
  responses:
    BadRequest:
      description: "请求格式无效或查询参数非法（invalid_json、invalid_request、invalid_input）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Unauthorized:
      description: "未登录或会话失效（authentication_required）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Forbidden:
      description: "被拒绝：跨站请求（cross_origin_request）、首次登录须改密（password_change_required）、需要管理员（administrator_required）或无权操作（forbidden）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    NotFound:
      description: "资源不存在（not_found）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Conflict:
      description: "版本或状态冲突（conflict、ledger_conflict、c2c_conflict、fee_rate_conflict 等）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Unprocessable:
      description: "业务校验失败（invalid_input、insufficient_spendable_capacity、unsafe_upstream 等）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
...
  schemas:
    Amount: {type: string, pattern: "^-?[0-9]+(\\.[0-9]{1,9})?$", description: "积分金额的十进制字符串，最多 9 位小数，不使用浮点数。"}
    Timestamp: {type: string, format: date-time, description: "RFC 3339 时间。"}
...
    Error:
      type: object
      properties:
        code: {type: string}
        message: {type: string}
      required: [code, message]
      additionalProperties: false
    ErrorResponse:
      type: object
      description: "所有 /api 错误的统一信封；code 为稳定的机器可读标识，message 为中文提示。"
      properties:
        error: {$ref: "#/components/schemas/Error"}
      required: [error]
      additionalProperties: false
...
    Wallet:
      type: object
      properties:
        posted_balance: {$ref: "#/components/schemas/Amount"}
        asset_reserved: {$ref: "#/components/schemas/Amount"}
        spend_authorized: {$ref: "#/components/schemas/Amount"}
        credit_limit: {$ref: "#/components/schemas/Amount"}
        effective_credit_limit: {$ref: "#/components/schemas/Amount"}
        credit_used: {$ref: "#/components/schemas/Amount"}
        credit_frozen: {type: boolean}
        spendable_capacity: {$ref: "#/components/schemas/Amount"}
        over_limit: {type: boolean}
        risk_status: {type: string, enum: [credit_frozen, over_limit, insufficient, normal]}
        updated_at: {$ref: "#/components/schemas/Timestamp"}
      required: [posted_balance, asset_reserved, spend_authorized, credit_limit, effective_credit_limit, credit_used, credit_frozen, spendable_capacity, over_limit, risk_status, updated_at]
      additionalProperties: false
    WalletEnvelope:
      type: object
      properties:
        wallet: {$ref: "#/components/schemas/Wallet"}
      required: [wallet]
      additionalProperties: false
    WalletResponse:
      type: object
      properties:
        wallet: {$ref: "#/components/schemas/Wallet"}
        recovery_actions:
          type: array
          items:
            type: object
            properties:
              kind: {type: string, enum: [market, my_orders]}
              href: {type: string}
            required: [kind, href]
            additionalProperties: false
      required: [wallet, recovery_actions]
      additionalProperties: false
    LedgerCounterparty:
      type: object
      properties:
        account_kind: {type: string, enum: [user, platform_incentive, platform_loss]}
        account_id: {type: string}
        business_role: {type: string}
        amount: {$ref: "#/components/schemas/Amount"}
      required: [account_kind, account_id, business_role, amount]
      additionalProperties: false
    LedgerEntry:
      type: object
      properties:
        id: {type: string, description: "账本分录 ID 的十进制字符串，可作为 before 游标。"}
        transaction_id: {type: string}
        entry_ordinal: {type: integer}
        business_role: {type: string}
        amount: {$ref: "#/components/schemas/Amount"}
        posted_balance_before: {$ref: "#/components/schemas/Amount"}
        posted_balance_after: {$ref: "#/components/schemas/Amount"}
        created_at: {$ref: "#/components/schemas/Timestamp"}
        transaction_kind: {type: string, enum: [transfer, admin_adjustment, bad_debt_transfer, hold_capture, self_channel_usage, reversal]}
        reason: {type: string}
        reference_type: {type: string}
        reference_id: {type: string}
        actor_account_id: {type: string}
        reversal_of_transaction_id: {type: string}
        hold_id: {type: string}
        counterparties:
          type: array
          items: {$ref: "#/components/schemas/LedgerCounterparty"}
      required: [id, transaction_id, entry_ordinal, business_role, amount, posted_balance_before, posted_balance_after, created_at, transaction_kind, reason, reference_type, reference_id, actor_account_id, reversal_of_transaction_id, hold_id, counterparties]
      additionalProperties: false
    LedgerTransactionEntry:
      type: object
      properties:
        id: {type: string, description: "账本分录 ID 的十进制字符串，可作为 before 游标。"}
        transaction_id: {type: string}
        entry_ordinal: {type: integer}
        business_role: {type: string}
        amount: {$ref: "#/components/schemas/Amount"}
        posted_balance_before: {$ref: "#/components/schemas/Amount"}
        posted_balance_after: {$ref: "#/components/schemas/Amount"}
        created_at: {$ref: "#/components/schemas/Timestamp"}
        transaction_kind: {type: string, enum: [transfer, admin_adjustment, bad_debt_transfer, hold_capture, self_channel_usage, reversal]}
        reason: {type: string}
        reference_type: {type: string}
        reference_id: {type: string}
        actor_account_id: {type: string}
        reversal_of_transaction_id: {type: string}
        hold_id: {type: string}
        counterparties:
          type: array
          items: {$ref: "#/components/schemas/LedgerCounterparty"}
        account_id: {type: string}
        account_kind: {type: string, enum: [user, platform_incentive, platform_loss]}
      required: [id, transaction_id, entry_ordinal, business_role, amount, posted_balance_before, posted_balance_after, created_at, transaction_kind, reason, reference_type, reference_id, actor_account_id, reversal_of_transaction_id, hold_id, counterparties, account_id, account_kind]
      additionalProperties: false
    LedgerEntryPage:
      type: object
      properties:
        entries:
          type: array
          items: {$ref: "#/components/schemas/LedgerEntry"}
        next_before: {type: string, description: "下一页游标；已无更多时为空字符串。"}
      required: [entries, next_before]
      additionalProperties: false
    LedgerTransaction:
      type: object
      properties:
        id: {type: string}
        idempotency_key: {type: string}
        kind: {type: string}
        reason: {type: string}
        reference_type: {type: string}
        reference_id: {type: string}
        actor_account_id: {type: string}
        reversal_of_transaction_id: {type: string}
        hold_id: {type: string}
        entries:
          type: array
          items: {$ref: "#/components/schemas/LedgerTransactionEntry"}
        created_at: {$ref: "#/components/schemas/Timestamp"}
      required: [id, idempotency_key, kind, reason, reference_type, reference_id, actor_account_id, reversal_of_transaction_id, hold_id, entries, created_at]
      additionalProperties: false
    LedgerTransactionEnvelope:
      type: object
      properties:
        transaction: {$ref: "#/components/schemas/LedgerTransaction"}
      required: [transaction]
      additionalProperties: false
    LedgerAccountRef:
      type: object
      description: "account_id 与 system_kind 二选一。"
      properties:
        account_id: {type: string, format: uuid}
        system_kind: {type: string, enum: [platform_incentive, platform_loss]}
      required: []
      additionalProperties: false
    LedgerAdjustmentRequest:
      type: object
      properties:
        from: {$ref: "#/components/schemas/LedgerAccountRef"}
        to: {$ref: "#/components/schemas/LedgerAccountRef"}
        amount: {$ref: "#/components/schemas/Amount"}
        reason: {type: string}
        reference_type: {type: string}
        reference_id: {type: string}
      required: [from, to, amount, reason]
      additionalProperties: false
    BadDebtRequest:
      type: object
      properties:
        account_id: {type: string, format: uuid}
        amount: {$ref: "#/components/schemas/Amount"}
        reason: {type: string}
        reference_id: {type: string}
      required: [account_id, amount, reason]
      additionalProperties: false
...
    C2CPaymentMethod:
      type: object
      properties:
        id: {type: string}
        type: {type: string, enum: [wechat, alipay, bank_transfer, other]}
        contact: {type: string}
        instructions: {type: string}
        qr_available: {type: boolean}
        qr_url: {type: string}
      required: [id, type, contact, instructions, qr_available, qr_url]
      additionalProperties: false
    C2COrderPaymentMethod:
      type: object
      properties:
        id: {type: string}
        type: {type: string, enum: [wechat, alipay, bank_transfer, other]}
        position: {type: integer}
        contact: {type: string}
        instructions: {type: string}
        qr_available: {type: boolean}
        qr_url: {type: string}
      required: [id, type, position, contact, instructions, qr_available, qr_url]
      additionalProperties: false
    C2COrder:
      type: object
      properties:
        id: {type: string}
        owner_account_id: {type: string}
        owner_display_name: {type: string}
        unit_price_fen: {type: integer}
        total: {$ref: "#/components/schemas/Amount"}
        available: {$ref: "#/components/schemas/Amount"}
        allocated: {$ref: "#/components/schemas/Amount"}
        settled: {$ref: "#/components/schemas/Amount"}
        closed: {$ref: "#/components/schemas/Amount"}
        minimum: {$ref: "#/components/schemas/Amount"}
        maximum: {$ref: "#/components/schemas/Amount"}
        status: {type: string, enum: [open, allocated, filled, cancelled]}
        takeable: {type: boolean}
        payment_types:
          type: array
          items: {type: string, enum: [wechat, alipay, bank_transfer, other]}
        payment_methods:
          type: array
          items: {$ref: "#/components/schemas/C2COrderPaymentMethod"}
        created_at: {$ref: "#/components/schemas/Timestamp"}
        updated_at: {$ref: "#/components/schemas/Timestamp"}
        cancelled_at:
          anyOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
      required: [id, owner_account_id, owner_display_name, unit_price_fen, total, available, allocated, settled, closed, minimum, maximum, status, takeable, payment_types, payment_methods, created_at, updated_at, cancelled_at]
      additionalProperties: false
...
    C2CTrade:
      type: object
      description: "参与方视图；不含管理员专属的处罚事件。"
      properties:
        id: {type: string}
        order_id: {type: string}
        buyer_account_id: {type: string}
        buyer_display_name: {type: string}
        seller_account_id: {type: string}
        seller_display_name: {type: string}
        quantity: {$ref: "#/components/schemas/Amount"}
        unit_price_fen: {type: integer}
        fiat_amount_fen: {type: integer}
        status: {type: string, enum: [awaiting_payment, paid, disputed, released_to_buyer, returned_to_seller, cancelled, expired]}
        payment_method:
          anyOf:
            - $ref: "#/components/schemas/C2CPaymentMethod"
            - type: "null"
        payment_reference: {type: string}
        payment_reference_deleted_at:
          anyOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        payment_deadline: {$ref: "#/components/schemas/Timestamp"}
        review_due_at:
          anyOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        ledger_transaction_id: {type: string}
        statements:
          type: array
          items:
            type: object
            properties:
              id: {type: string}
              actor_account_id: {type: string}
              actor_display_name: {type: string}
              text: {type: string}
              character_count: {type: integer}
              created_at: {$ref: "#/components/schemas/Timestamp"}
              deleted_at:
                anyOf:
                  - $ref: "#/components/schemas/Timestamp"
                  - type: "null"
            required: [id, actor_account_id, actor_display_name, text, character_count, created_at, deleted_at]
            additionalProperties: false
        events:
          type: array
          items:
            type: object
            properties:
              id: {type: integer}
              actor_account_id: {type: string}
              action: {type: string}
              reason: {type: string}
              ledger_transaction_id: {type: string}
              created_at: {$ref: "#/components/schemas/Timestamp"}
            required: [id, actor_account_id, action, reason, ledger_transaction_id, created_at]
            additionalProperties: false
        created_at: {$ref: "#/components/schemas/Timestamp"}
        updated_at: {$ref: "#/components/schemas/Timestamp"}
        paid_at:
          anyOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        resolved_at:
          anyOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
      required: [id, order_id, buyer_account_id, buyer_display_name, seller_account_id, seller_display_name, quantity, unit_price_fen, fiat_amount_fen, status, payment_method, payment_reference, payment_reference_deleted_at, payment_deadline, review_due_at, ledger_transaction_id, statements, events, created_at, updated_at, paid_at, resolved_at]
      additionalProperties: false
...
    C2CMyActivity:
      type: object
      properties:
        orders:
          type: array
          items: {$ref: "#/components/schemas/C2COrder"}
        trades:
          type: array
          items: {$ref: "#/components/schemas/C2CTrade"}
      required: [orders, trades]
      additionalProperties: false
...